# api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Assisted Migration Agent API
  version: v2
servers:
  - url: /api/v2
...
paths:
...
  /collections/{id}/virtualmachines:
    get:
      tags: [VirtualMachines]
      summary: List VirtualMachines in a collection
      operationId: listVirtualMachines
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: byExpression
          in: query
          description: Filter by expression
          schema:
            type: string
        - name: sort
          in: query
          description: Sort fields with direction (e.g., "name:asc", "cluster:desc"). Valid fields are name, vCenterState, cluster, diskSize, memory, issues, cpuUsage, diskUsage, ramUsage, cpuAvg, memAvg.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: ["cluster:asc", "name:desc"]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: List of VirtualMachines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualMachineListResponse'
        '400':
          description: Invalid request parameters
        '404':
          description: Collection not found
        '500':
          description: Internal server error
  /collections/{id}/virtualmachines/batch-update-exclusion:
    post:
      tags: [VirtualMachines]
      summary: Batch update migration exclusion for multiple VMs
      operationId: batchUpdateVMExclusion
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateExclusionRequest'
      responses:
        '200':
          description: Exclusion updated
        '400':
          description: Invalid request
        '404':
          description: Collection or VM not found
        '500':
          description: Internal server error
  /collections/{id}/virtualmachines/filter-options:
    get:
      tags: [VirtualMachines]
      summary: Get available filter option values
      operationId: getVMFilterOptions
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Filter options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMFilterOptionsResponse'
        '404':
          description: Collection not found
        '500':
          description: Internal server error
  /collections/{id}/virtualmachines/labels:
    get:
      tags: [VirtualMachines]
      summary: Get all distinct labels with counts
      operationId: getVMLabels
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Labels with counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMLabelsResponse'
        '404':
          description: Collection not found
        '500':
          description: Internal server error
  /collections/{id}/virtualmachines/labels/{label}:
    patch:
      tags: [VirtualMachines]
      summary: Add or remove a label from multiple VMs
      operationId: updateLabelVMs
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: label
          in: path
          required: true
          description: Label name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLabelVMsRequest'
      responses:
        '200':
          description: Label updated
        '400':
          description: Invalid request
        '404':
          description: Collection or VM not found
        '500':
          description: Internal server error
    delete:
      tags: [VirtualMachines]
      summary: Remove a label from all VMs
      operationId: deleteLabelGlobally
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: label
          in: path
          required: true
          description: Label name
          schema:
            type: string
      responses:
        '200':
          description: Label removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteLabelGloballyResponse'
        '400':
          description: Invalid request
        '500':
          description: Internal server error
  /collections/{id}/virtualmachines/{vmId}:
    get:
      tags: [VirtualMachines]
      summary: Get VirtualMachine details
      operationId: getVirtualMachine
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: vmId
          in: path
          required: true
          description: VirtualMachine ID
          schema:
            type: string
      responses:
        '200':
          description: VirtualMachine details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualMachineDetail'
        '404':
          description: Collection or VirtualMachine not found
        '500':
          description: Internal server error
    patch:
      tags: [VirtualMachines]
      summary: Update VirtualMachine properties
      operationId: updateVirtualMachine
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: vmId
          in: path
          required: true
          description: VirtualMachine ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VirtualMachineUpdateRequest'
      responses:
        '200':
          description: VirtualMachine updated
        '400':
          description: Invalid request
        '404':
          description: Collection or VirtualMachine not found
        '500':
          description: Internal server error

  # ── Export ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    VirtualMachine:
      type: object
      required:
        - name
        - id
        - vCenterID
        - vCenterState
        - cluster
        - datacenter
        - diskSize
        - memory
        - issueCount
      properties:
        name:
          type: string
          description: VirtualMachine name
        id:
          type: string
          description: VirtualMachine ID in vCenter
        vCenterID:
          type: string
          description: Identifier of the vCenter this VirtualMachine belongs to
        vCenterState:
          type: string
          description: vCenter state (e.g., poweredOn, poweredOff, suspended)
        cluster:
          type: string
          description: Cluster name
        datacenter:
          type: string
          description: Datacenter name
        diskSize:
          type: integer
          format: int64
          description: Total disk size in MB
        memory:
          type: integer
          format: int64
          description: Memory size in MB
        issueCount:
          type: integer
          description: Number of issues found for this VirtualMachine
        migratable:
          type: boolean
          description: True if the VirtualMachine is migratable for MTV
        template:
          type: boolean
          description: True if the VirtualMachine is a template
        inspectionStatus:
          $ref: '#/components/schemas/InspectionStatus'
        inspectionConcernCount:
          type: integer
          description: Number of inspection concerns for the latest persisted result
        tags:
          type: array
          items:
            type: string
          description: Tags aggregated from matching groups
    VirtualMachineDetail:
      type: object
      required:
        - id
        - name
        - vCenterID
        - powerState
        - connectionState
        - cpuCount
        - coresPerSocket
        - memoryMB
        - disks
        - nics
      properties:
        id:
          type: string
          description: Unique identifier for the VirtualMachine in vCenter
        name:
          type: string
          description: Display name of the VirtualMachine
        vCenterID:
          type: string
          description: Identifier of the vCenter this VirtualMachine belongs to
        uuid:
          type: string
          description: Universally unique identifier assigned by vCenter
        firmware:
          type: string
          description: Firmware type (bios or efi)
        powerState:
          type: string
          description: Current power state (poweredOn, poweredOff, suspended)
        connectionState:
          type: string
          description: Connection state (connected, disconnected, orphaned, inaccessible)
        host:
          type: string
          description: ESXi host reference
        datacenter:
          type: string
          description: Datacenter name
        cluster:
          type: string
          description: Cluster name
        folder:
          type: string
          description: Inventory folder reference
        cpuCount:
          type: integer
          format: int32
          description: Total virtual CPUs
        coresPerSocket:
          type: integer
          format: int32
          description: CPU cores per virtual socket
        cpuAffinity:
          type: array
          items:
            type: integer
            format: int32
          description: Physical CPU IDs the VirtualMachine is pinned to
        memoryMB:
          type: integer
          format: int32
          description: Memory in megabytes
        guestName:
          type: string
          description: Guest OS full name
        guestId:
          type: string
          description: VMware guest OS identifier
        hostName:
          type: string
          description: Guest hostname
        ipAddress:
          type: string
          description: Guest primary IP address
        storageUsed:
          type: integer
          format: int64
          description: Storage consumed in bytes
        template:
          type: boolean
          description: Whether the VirtualMachine is a template
        migratable:
          type: boolean
          description: Whether the VirtualMachine can be migrated
        faultToleranceEnabled:
          type: boolean
          description: Whether Fault Tolerance is enabled
        nestedHVEnabled:
          type: boolean
          description: Whether nested virtualization is enabled
        toolsStatus:
          type: string
          description: VMware Tools installation status
        toolsRunningStatus:
          type: string
          description: Whether VMware Tools is running
        disks:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachineDisk'
        nics:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachineNIC'
        devices:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachineDevice'
        guestNetworks:
          type: array
          items:
            $ref: '#/components/schemas/GuestNetwork'
        issues:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachineIssue'
        inspection:
          $ref: '#/components/schemas/VirtualMachineInspectionResults'
    VirtualMachineListResponse:
      type: object
      required:
        - virtualMachines
        - total
        - page
        - pageCount
      properties:
        virtualMachines:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachine'
        total:
          type: integer
          description: Total number of VirtualMachines matching the filter
        page:
          type: integer
          description: Current page number
        pageCount:
          type: integer
          description: Total number of pages

    # ── VirtualMachine Sub-resources ─────────────────────────────────────────────────
    VirtualMachineDisk:
      type: object
      properties:
        key:
          type: integer
          format: int32
          description: Unique disk key within the VirtualMachine
        file:
          type: string
          description: VMDK file path in the datastore
        capacity:
          type: integer
          format: int64
          description: Disk capacity in bytes
        shared:
          type: boolean
          description: Whether the disk is shared between multiple VirtualMachines
        rdm:
          type: boolean
          description: Whether this is a Raw Device Mapping
        bus:
          type: string
          description: Bus type (scsi, ide, sata, nvme)
        mode:
          type: string
          description: Disk mode (persistent, independent_persistent, independent_nonpersistent)
    VirtualMachineNIC:
      type: object
      properties:
        mac:
          type: string
          description: MAC address
        network:
          type: string
          description: Network reference
        index:
          type: integer
          description: NIC index within the VirtualMachine
    VirtualMachineDevice:
      type: object
      properties:
        kind:
          type: string
          description: Device type (cdrom, floppy, usb, serial, parallel)
    GuestNetwork:
      type: object
      properties:
        device:
          type: string
          description: Network device name inside the guest
        mac:
          type: string
          description: MAC address as seen by the guest
        ip:
          type: string
          description: IP address
        prefixLength:
          type: integer
          format: int32
          description: Network prefix length (CIDR)
        network:
          type: string
          description: Network name from the guest
    VirtualMachineIssue:
      type: object
      required:
        - label
        - category
        - description
      properties:
        label:
          type: string
          description: Short label
        description:
          type: string
          description: Detailed description with context and recommendations
        category:
          type: string
          enum:
            - Critical
            - Warning
            - Information
            - Advisory
            - Error
            - Other
          x-enum-varnames:
            - VirtualMachineIssueCategoryCritical
            - VirtualMachineIssueCategoryWarning
            - VirtualMachineIssueCategoryInformation
            - VirtualMachineIssueCategoryAdvisory
            - VirtualMachineIssueCategoryError
            - VirtualMachineIssueCategoryOther

    # ── Inspection ───────────────────────────────────────────────────────
    InspectionStatus:
      type: object
      required:
        - state
      properties:
        state:
          type: string
          enum:
            - pending
            - running
            - completed
            - canceled
            - error
          x-enum-varnames:
            - InspectionStatusStatePending
            - InspectionStatusStateRunning
            - InspectionStatusStateCompleted
            - InspectionStatusStateCanceled
            - InspectionStatusStateError
          description: Current inspection state
        error:
          type: string
          description: Error message when state is error
    VirtualMachineInspectionResults:
      type: object
      description: VirtualMachines Inspection results
      properties:
        concerns:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachineInspectionConcern'
    VirtualMachineInspectionConcern:
      type: object
      required:
        - message
        - category
        - label
      properties:
        label:
          type: string
          description: Short label identifying the concern
        message:
          type: string
        category:
          type: string

    # ── Groups ───────────────────────────────────────────────────────────
...
    VirtualMachineUpdateRequest:
      type: object
      properties:
        migrationExcluded:
          type: boolean
          description: Whether to exclude this VM from migration
          x-oapi-codegen-extra-tags:
            binding: "omitempty"
        labels:
          type: array
          items:
            type: string
          description: User-defined labels (replaces existing labels)
          x-oapi-codegen-extra-tags:
            binding: "omitempty,dive,notblank,min=1,max=100"
    BatchUpdateExclusionRequest:
      type: object
      required:
        - vmIds
        - migrationExcluded
      properties:
        vmIds:
          type: array
          items:
            type: string
          description: VM IDs to update
          x-oapi-codegen-extra-tags:
            binding: "required,min=1,dive,required"
        migrationExcluded:
          type: boolean
          description: Exclusion state to set for all VMs
    VMFilterOptionsResponse:
      type: object
      required:
        - clusters
        - datacenters
        - concernLabels
        - concernCategories
        - applications
      properties:
        clusters:
          type: array
          items:
            type: string
          description: Distinct cluster names
        datacenters:
          type: array
          items:
            type: string
          description: Distinct datacenter names
        concernLabels:
          type: array
          items:
            type: string
          description: Distinct concern labels
        concernCategories:
          type: array
          items:
            type: string
          description: Distinct concern categories
        applications:
          type: array
          items:
            type: string
          description: Distinct detected application names
    VMLabelsResponse:
      type: object
      required:
        - labels
        - counts
      properties:
        labels:
          type: array
          items:
            type: string
          description: Distinct labels currently in use
        counts:
          type: array
          items:
            type: integer
          description: Number of VMs with each label (same order as labels array)
    UpdateLabelVMsRequest:
      type: object
      properties:
        add:
          type: array
          items:
            type: string
          description: VMs to add the label to
          x-oapi-codegen-extra-tags:
            binding: "omitempty,dive,required"
        remove:
          type: array
          items:
            type: string
          description: VMs to remove the label from
          x-oapi-codegen-extra-tags:
            binding: "omitempty,dive,required"
    DeleteLabelGloballyResponse:
      type: object
      required:
        - label
        - affected
      properties:
        label:
          type: string
          description: The label that was removed
        affected:
          type: integer
          description: Number of VMs that had the label removed

    # ── Agent ────────────────────────────────────────────────────────────
...